devops/10-sre-operacoes/requirements.txt
# Dependências do Módulo 10 — SRE e Operações
# Instalar com: pip install -r requirements.txt

# App de exemplo (se precisar demo)
fastapi==0.115.4
uvicorn[standard]==0.32.0

# Suporte para scripts didaticos
pyyaml==6.0.2
rich==13.9.3
requests==2.32.3
python-dateutil==2.9.0.post0
kubernetes==31.0.0

# Testes
pytest==8.3.3
